+++ src/OnionCrafter.Feature/Handlers/General/Command/BaseRequestCommandHandler.cs
@@ -1,32 +1,33 @@
 ﻿using AutoMapper;
 using OnionCrafter.Base.DTOs;
 using OnionCrafter.Base.Wrappers.Requests;
+using OnionCrafter.Base.Wrappers.Requests.Command;
 using OnionCrafter.Base.Wrappers.Responses;
 
-namespace OnionCrafter.Feature.Handlers
+namespace OnionCrafter.Feature.Handlers.General.Command
 {
     /// <summary>
-    /// Represents a base class for query request handlers.
+    /// Represents a base class for command request handlers.
     /// </summary>
     /// <typeparam name="TSchemaKey">The type of the schema key.</typeparam>
     /// <typeparam name="TRequestSchema">The type of the request schema.</typeparam>
     /// <typeparam name="TRequestDTO">The type of the request DTO.</typeparam>
     /// <typeparam name="TResponseSchema">The type of the response schema.</typeparam>
     /// <typeparam name="TReturnDTO">The type of the return DTO.</typeparam>
-    public abstract class BaseRequestQueryHandler<TSchemaKey, TRequestSchema, TRequestDTO, TResponseSchema, TReturnDTO> :
+    public abstract class BaseRequestCommandHandler<TSchemaKey, TRequestSchema, TRequestDTO, TResponseSchema, TReturnDTO> :
         BaseRequestHandler<TSchemaKey, TRequestSchema, TRequestDTO, TResponseSchema, TReturnDTO>,
-        IRequestQueryHandler<TSchemaKey, TRequestSchema, TRequestDTO, TResponseSchema, TReturnDTO>
+        IRequestCommandHandler<TSchemaKey, TRequestSchema, TRequestDTO, TResponseSchema, TReturnDTO>
         where TSchemaKey : notnull, IEquatable<TSchemaKey>, IComparable<TSchemaKey>
         where TResponseSchema : IResponseSchema<TSchemaKey, TReturnDTO>
-        where TRequestSchema : IRequestSchema<TSchemaKey, TResponseSchema, TReturnDTO, TRequestDTO>
+        where TRequestSchema : ICommandRequestSchema<TSchemaKey, TResponseSchema, TReturnDTO, TRequestDTO>
         where TReturnDTO : class, IBaseDTO, IResponseData
         where TRequestDTO : class, IBaseDTO, IRequestData
     {
         /// <summary>
-        /// Initializes a new instance of the <see cref="BaseRequestQueryHandler{TSchemaKey, TRequestSchema, TRequestDTO, TResponseSchema, TReturnDTO}"/> class.
+        /// Initializes a new instance of the <see cref="BaseRequestCommandHandler{TSchemaKey, TRequestSchema, TRequestDTO, TResponseSchema, TReturnDTO}"/> class.
         /// </summary>
         /// <param name="mapper">The object mapper instance.</param>
-        protected BaseRequestQueryHandler(IMapper mapper) : base(mapper)
+        protected BaseRequestCommandHandler(IMapper mapper) : base(mapper)
         {
         }
     }
