+++ src/OnionCrafter.Feature/Handlers/General/IRequestHandler.cs
@@ -3,7 +3,7 @@ using OnionCrafter.Base.DTOs;
 using OnionCrafter.Base.Wrappers.Requests;
 using OnionCrafter.Base.Wrappers.Responses;
 
-namespace OnionCrafter.Feature.Handlers
+namespace OnionCrafter.Feature.Handlers.General
 {
     /// <summary>
     /// Represents a base interface for request handlers.
@@ -14,12 +14,12 @@ namespace OnionCrafter.Feature.Handlers
     /// <typeparam name="TResponseSchema">The type of the response schema.</typeparam>
     /// <typeparam name="TReturnDTO">The type of the return DTO.</typeparam>
     public interface IRequestHandler<TSchemaKey, TRequestSchema, TRequestDTO, TResponseSchema, TReturnDTO> :
-        IRequestHandler<TRequestSchema, TResponseSchema>
+        IBaseRequestHandler<TRequestSchema, TResponseSchema>
         where TSchemaKey : notnull, IEquatable<TSchemaKey>, IComparable<TSchemaKey>
-        where TRequestSchema : IRequestSchema<TSchemaKey, TResponseSchema, TReturnDTO, TRequestDTO>
         where TRequestDTO : class, IBaseDTO, IRequestData
-        where TResponseSchema : IResponseSchema<TSchemaKey, TReturnDTO>
         where TReturnDTO : class, IBaseDTO, IResponseData
+        where TRequestSchema : IBaseRequestSchema, IRequest<TResponseSchema>
+        where TResponseSchema : IBaseResponseSchema
     {
     }
 }
