--- src/OnionCrafter.Feature/Handlers/IBaseRequestHandler.cs
@@ -1,9 +1,0 @@
-﻿namespace OnionCrafter.Feature.Handlers
-{
-    /// <summary>
-    /// Interface representing a base request handler
-    /// </summary>
-    public interface IBaseRequestHandler
-    {
-    }
-}
