+++ src/OnionCrafter.Feature/Handlers/Services/BaseRequestServiceHandler.cs
@@ -2,8 +2,9 @@
 using OnionCrafter.Base.DTOs;
 using OnionCrafter.Base.Wrappers.Requests;
 using OnionCrafter.Base.Wrappers.Responses;
+using OnionCrafter.Feature.Handlers.General;
+using OnionCrafter.Service.Options.Services;
 using OnionCrafter.Service.Services;
-using OnionCrafter.Service.Services.Options;
 
 namespace OnionCrafter.Feature.Handlers.Services
 {
@@ -61,7 +62,7 @@ namespace OnionCrafter.Feature.Handlers.Services
         where TReturnDTO : class, IBaseDTO, IResponseData
         where TRequestDTO : class, IBaseDTO, IRequestData
         where TService : IBaseService
-        where TServiceOptions : class, IServiceOptions
+        where TServiceOptions : class, IBaseServiceOptions
     {
         /// <summary>
         /// The service options instance.
