+++ src/OnionCrafter.Feature/Handlers/Services/Command/BaseRequestServiceCommandHandler.cs
@@ -2,10 +2,10 @@
 using OnionCrafter.Base.DTOs;
 using OnionCrafter.Base.Wrappers.Requests;
 using OnionCrafter.Base.Wrappers.Responses;
+using OnionCrafter.Service.Options.Services;
 using OnionCrafter.Service.Services;
-using OnionCrafter.Service.Services.Options;
 
-namespace OnionCrafter.Feature.Handlers.Services
+namespace OnionCrafter.Feature.Handlers.Services.Command
 {
     /// <summary>
     /// Represents a base class for service command handlers.
@@ -55,7 +55,7 @@ namespace OnionCrafter.Feature.Handlers.Services
         where TReturnDTO : class, IBaseDTO, IResponseData
         where TRequestDTO : class, IBaseDTO, IRequestData
         where TService : IBaseService
-        where TServiceOptions : class, IServiceOptions
+        where TServiceOptions : class, IBaseServiceOptions
     {
         /// <summary>
         /// Initializes a new instance of the <see cref="BaseRequestServiceCommandHandler{TSchemaKey, TRequestSchema, TRequestDTO, TResponseSchema, TReturnDTO, TService, TServiceOptions}"/> class.
