+++ src/OnionCrafter.Feature/Handlers/Services/IRequestServiceHandler.cs
@@ -1,7 +1,8 @@
 ﻿using OnionCrafter.Base.DTOs;
 using OnionCrafter.Base.Wrappers.Requests;
 using OnionCrafter.Base.Wrappers.Responses;
-using OnionCrafter.Service.Services.Options;
+using OnionCrafter.Feature.Handlers.General;
+using OnionCrafter.Service.Options.Services;
 using OnionCrafter.Service.Services;
 
 namespace OnionCrafter.Feature.Handlers.Services
@@ -15,7 +16,9 @@ namespace OnionCrafter.Feature.Handlers.Services
     /// <typeparam name="TResponseSchema">The type of the response schema.</typeparam>
     /// <typeparam name="TReturnDTO">The type of the return DTO.</typeparam>
     /// <typeparam name="TService">The type of the service.</typeparam>
-    public interface IRequestServiceHandler<TSchemaKey, TRequestSchema, TRequestDTO, TResponseSchema, TReturnDTO, TService> : IRequestHandler<TSchemaKey, TRequestSchema, TRequestDTO, TResponseSchema, TReturnDTO>
+    public interface IRequestServiceHandler<TSchemaKey, TRequestSchema, TRequestDTO, TResponseSchema, TReturnDTO, TService> :
+        IRequestHandler<TSchemaKey, TRequestSchema, TRequestDTO, TResponseSchema, TReturnDTO>,
+        IBaseRequestServiceHandler<TRequestSchema, TResponseSchema>
         where TSchemaKey : notnull, IEquatable<TSchemaKey>, IComparable<TSchemaKey>
         where TResponseSchema : IResponseSchema<TSchemaKey, TReturnDTO>
         where TRequestSchema : IRequestSchema<TSchemaKey, TResponseSchema, TReturnDTO, TRequestDTO>
@@ -35,14 +38,15 @@ namespace OnionCrafter.Feature.Handlers.Services
     /// <typeparam name="TReturnDTO">The type of the return DTO.</typeparam>
     /// <typeparam name="TService">The type of the service.</typeparam>
     /// <typeparam name="TServiceOptions">The type of the service options.</typeparam>
-    public interface IRequestServiceHandler<TSchemaKey, TRequestSchema, TRequestDTO, TResponseSchema, TReturnDTO, TService, TServiceOptions> : IRequestServiceHandler<TSchemaKey, TRequestSchema, TRequestDTO, TResponseSchema, TReturnDTO, TService>
+    public interface IRequestServiceHandler<TSchemaKey, TRequestSchema, TRequestDTO, TResponseSchema, TReturnDTO, TService, TServiceOptions> :
+        IRequestServiceHandler<TSchemaKey, TRequestSchema, TRequestDTO, TResponseSchema, TReturnDTO, TService>
         where TSchemaKey : notnull, IEquatable<TSchemaKey>, IComparable<TSchemaKey>
         where TResponseSchema : IResponseSchema<TSchemaKey, TReturnDTO>
         where TRequestSchema : IRequestSchema<TSchemaKey, TResponseSchema, TReturnDTO, TRequestDTO>
         where TReturnDTO : class, IBaseDTO, IResponseData
         where TRequestDTO : class, IBaseDTO, IRequestData
         where TService : IBaseService
-        where TServiceOptions : class, IServiceOptions
+        where TServiceOptions : class, IBaseServiceOptions
     {
     }
 }
